Export OcTree (#417)
* Update exprts

* (transpiler) read revision correctly

* Fix typo

* Make transpiler more robust

More robust to whitespace; fix dot bug (dots are any character in regex); Arguments are optional to constructors

* OccupancyMap* -> OcTree*

Rename as to similar to OccupancyGrid*; Including seperation of all classes to their own file, needed by transpiler; Remove all seperate class property definitions; not supported by transpiler
