+++ es6-support/index.js
@@ -23,6 +23,9 @@ export * from './models/TriangleList'
 
 export * from './navigation/OccupancyGrid'
 export * from './navigation/OccupancyGridClient'
+export * from './navigation/OcTree'
+export * from './navigation/ColorOcTree'
+export * from './navigation/OcTreeClient'
 export * from './navigation/Odometry'
 export * from './navigation/Path'
 export * from './navigation/Point'
